perf: extract Uptime component to isolate 1Hz re-renders from Sidebar
